homework24/volkau_app/requirements.txt
asgiref==3.6.0
asttokens==2.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.5
django-allauth==0.52.0
django-bootstrap-v5==1.0.11
django-ckeditor==6.5.1
django-crispy-forms==2.0
django-filter==22.1
django-js-asset==2.0.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
drf-yasg==1.21.5
entrypoints==0.4
executing==1.2.0
idna==3.4
inflection==0.5.1
ipykernel==6.16.2
ipython==8.5.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jupyter_client==7.4.4
jupyter_core==4.11.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
oauthlib==3.2.2
packaging==21.3
parso==0.8.3
peewee==3.15.4
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.31
psutil==5.9.3
psycopg2==2.9.5
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2022.7.1
pywin32==304
pyzmq==24.0.1
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.0
tornado==6.2
traitlets==5.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
wcwidth==0.2.5
